* dln.c (init_funcname_len, dln_find_exe_r, dln_find_file_r): use   size_t.
* file.c (rb_stat_inspect, file_expand_path): ditto.

* util.c (ruby_qsort): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@22944 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ include/ruby/util.h
@@ -49,7 +49,7 @@ unsigned long ruby_scan_hex(const char *, int, int *);
 void ruby_add_suffix(VALUE str, const char *suffix);
 #endif
 
-void ruby_qsort(void *, const int, const int,
+void ruby_qsort(void *, const size_t, const size_t,
 		int (*)(const void *, const void *, void *), void *);
 
 void ruby_setenv(const char *, const char *);
